mcp-gateway/pyproject.toml
[project]
name = "aviary-mcp-gateway"
version = "0.1.0"
description = "Aviary MCP Gateway — tool catalog + proxy"
requires-python = ">=3.12"
dependencies = [
    "aviary-shared",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
